bakery/src/scripts/requirements.txt
PyICU==1.9.8
cnx-epub==0.21.0
cnx-common==1.2.2
cnx-easybake==1.2.7
awscli==1.18.39
boto3==1.12.39
python-magic==0.4.15